Sesam

Documentation for Sesam.

Sesam.SystemParUpdater

Sesam.calculate_β_NR_sesam3

Sesam.plant_face

Sesam.SystemParUpdater — Type

SystemParUpdater(d::Dict, sys::ModelingToolkit.AbstractSystem)

helps updating a subset of a parameter vector by another vector. The positions to update are specified during initialization by providing the names or a dictionary. The Callable can then be used to construct an updated Problem without recreating the Problem from the System.

# given an ODESystem, sys, and associated ODEProblem, prob:

d_up = Dict(p3 => 0.3, p1 => 0.1) # a subset of the original parameters
upd! = SystemParUpdater(d_up,sys)
upd!(odeproblem_of_sys.p, d_up)

Sesam.calculate_β_NR_sesam3 — Method

R pool is a mixture of microbial turnover and enzymes. Here the stable C/N ratio is computed based on given parameterization.

Sesam.plant_face — Method

model iL and βNi as as step function increased by factor fac_int at t in [t1,t2)