Sesam

Documentation for Sesam.

Sesam.calculate_β_NR_sesam3

Sesam.plant_face

Sesam.plant_face_fluct

Sesam.smoothstep

Sesam.calculate_β_NR_sesam3 — Method

R pool is a mixture of microbial turnover and enzymes. Here the stable C/N ratio is computed based on given parameterization.

Sesam.plant_face — Method

model iL and βNi as as step function increased by factor fac_int at t in [t1,t2)

Sesam.plant_face_fluct — Method

model iL and βNi as as step function increased by factor fac_int at t in [t1,t2). Superimpose a seasonal pattern on the litter input.

Sesam.smoothstep — Function

smoothstep(x, x_step, dx, a=zero(x), b=one(x))

smooth step function:

= a for x <= x_step - dx

= b for x >= x_step + dx

= (a-b)/2 for x == x_step

and smooth in between.