Sesam

Documentation for Sesam.

Sesam.calculate_propR

Sesam.calculate_shareEnzTvr

Sesam.calculate_β_NBtvr

Sesam.calculate_β_NR_sesam3

Sesam.calculate_β_PR_sesam3

Sesam.compute_elemental_weightfactor

Sesam.compute_mean_du2

Sesam.compute_mean_du3

Sesam.get_updated_Penz_pars

Sesam.plant_const

Sesam.plant_const_balanced

Sesam.plant_face

Sesam.plant_face_fluct

Sesam.sesam_const_dLRP

Sesam.calculate_propR — Method

Given the C/N ratio of total SOM, litter input and input to R: tvr of B and Enz calculate the proportion of residue pool R to total SOM

Sesam.calculate_shareEnzTvr — Method

Calculate the share of enzymes tvr contributing to total tvr in addition to biomass

Sesam.calculate_β_NBtvr — Method

Calculate the C/N ratio of microbial biomass turnover. It may differ from β_NB because of elemental resorption.

Sesam.calculate_β_NR_sesam3 — Method

R pool is a mixture of microbial turnover and enzymes. Here the stable C/N ratio is computed based on given parameterization.

Sesam.calculate_β_PR_sesam3 — Method

R pool is a mixture of microbial turnover and enzymes. Here the stable C/N ratio is computed based on given parameterization.

Sesam.compute_elemental_weightfactor — Function

compute elemental-limitation weighting factor omega_Z

Sesam.compute_mean_du2 — Method

compute_mean_du2(du1, α1, du2, α2; sum_α_others=0)

Compute mean across derivates, du, with checking excluding enzymes for 2 enzymes Given derivates du1, du2 and allocations α1 and α2. (Wutzler22 Appendix D1)

sumαothers: allocation to enzymes that should be excluded from the mean

Checks all the combinations of relative change of du being smaller than -α_Z

Sesam.compute_mean_du3 — Method

compute_mean_du3(du1, α1, du2, α2, du3, α3; sum_α_others=0)

Compute mean across derivates, du, with checking excluding enzymes for 2 enzymes Given derivates du1, du2, du3 and allocations α1, α2, α3. (Wutzler22 Appendix D1)

sumαothers: allocation to enzymes that should be excluded from the mean

Checks all the combinations of relative change of du being smaller than -α_Z

Sesam.get_updated_Penz_pars — Method

get_updated_Penz_pars(pDict, s)

Create a copy of parameters and set kmNP, kLP, and kRP to values of the L/R enzymes.

Sesam.plant_const — Method

System to be coupled with Sesam3 that provides constant inputs, i.e. drivers. Need to specify parameters

i_L0: litter input rate

β_Ni0, β_Pi0: C:N and C:P ratio of litter input

i_IN0, i_IP0: external inputs of inorganic N and P (N deposition, weathering)

optional:

u_PlantNmax0, u_PlantPmax0: maximum plant uptake of inorganic N and P

k_PlantN0, k_PlantP0: plant uptake rate per inorganic pool of N and P

u_Pup: plant inorganic P uplift (to potentially modify u_PlantPmax0)

Defaults of plant uptake are set so that uptake rate is very high (1e3) so that the maximum is taken up, i.e. wins competition with microbes. The maximum plan uptake is set to match the input by litter.

Sesam.plant_const_balanced — Method

System to be coupled with Sesam3 that provides constant inputs, i.e. drivers. Plant inorganic N and P uptake are computed to balance litter fall. Need to specify parameters

i_L0: litter input rate

β_Ni0, β_Pi0: C:N and C:P ratio of litter input

i_IN0, i_IP0: external inputs of inorganic N and P (N deposition, weathering)

optional:

imbalance_N0, imbalance_P0: to compute u_PlantNmax = iL0 / β_Ni0 - imbalance_N: negative for N or P exported or sequestered in biomass and positive for N or P taken up, e.g. from non-modelled deeper soil layers.

k_PlantN0, k_PlantP0: default to very high value, so that u_PlantNmax0 are realized.

Defaults of plant uptake are set so that uptake rate is very high (1e3) so that the maximum is taken up, i.e. wins competition with microbes. The maximum plan uptake is set to match the input by litter.

Sesam.plant_face — Method

model iL and βNi as as step function increased by factor fac_int at t in [t1,t2)

Sesam.plant_face_fluct — Method

model iL and βNi as as step function increased by factor fac_int at t in [t1,t2). Superimpose a seasonal pattern on the litter input.

Sesam.sesam_const_dLRP — Method

System where dL, dR, and d_P are fixed to given parameters, B held constant (dB/dt = 0)