Sesam

Documentation for Sesam.

Sesam.calculate_β_NR_sesam3

Sesam.calculate_β_PR_sesam3

Sesam.plant_const

Sesam.plant_const_balanced

Sesam.plant_face

Sesam.plant_face_fluct

Sesam.calculate_β_NR_sesam3 — Method

R pool is a mixture of microbial turnover and enzymes. Here the stable C/N ratio is computed based on given parameterization.

Sesam.calculate_β_PR_sesam3 — Method

R pool is a mixture of microbial turnover and enzymes. Here the stable C/N ratio is computed based on given parameterization.

Sesam.plant_const — Method

System to be coupled with Sesam3 that provides constant inputs, i.e. drivers. Need to specify parameters

i_L0: litter input rate

β_Ni0, β_Pi0: C:N and C:P ratio of litter input

i_IN0, i_IP0: external inputs of inorganic N and P (N deposition, weathering)

optional:

u_PlantNmax0, u_PlantPmax0: maximum plant uptake of inorganic N and P

k_PlantN0, k_PlantP0: plant uptake rate per inorganic pool of N and P

u_Pup: plant inorganic P uplift (to potentially modify u_PlantPmax0)

Defaults of plant uptake are set so that uptake rate is very high (1e3) so that the maximum is taken up, i.e. wins competition with microbes. The maximum plan uptake is set to match the input by litter.

Sesam.plant_const_balanced — Method

System to be coupled with Sesam3 that provides constant inputs, i.e. drivers. Plant inorganic N and P uptake are computed to balance litter fall. Need to specify parameters

i_L0: litter input rate

β_Ni0, β_Pi0: C:N and C:P ratio of litter input

i_IN0, i_IP0: external inputs of inorganic N and P (N deposition, weathering)

optional:

imbalance_N0, imbalance_P0: to compute u_PlantNmax = iL0 / β_Ni0 - imbalance_N: negative for N or P exported or sequestered in biomass and positive for N or P taken up, e.g. from non-modelled deeper soil layers.

k_PlantN0, k_PlantP0: default to very high value, so that u_PlantNmax0 are realized.

Defaults of plant uptake are set so that uptake rate is very high (1e3) so that the maximum is taken up, i.e. wins competition with microbes. The maximum plan uptake is set to match the input by litter.

Sesam.plant_face — Method

model iL and βNi as as step function increased by factor fac_int at t in [t1,t2)

Sesam.plant_face_fluct — Method

model iL and βNi as as step function increased by factor fac_int at t in [t1,t2). Superimpose a seasonal pattern on the litter input.